fix mute on channels
