+++ src/Game/NeoServer.Game.Contracts/Chats/IChatChannel.cs
@@ -14,7 +14,7 @@ namespace NeoServer.Game.Contracts.Chats
     {
         ushort Id { get; }
         string Name { get; }
-        IEnumerable<IChatUser> Users { get; }
+        IEnumerable<IUserChat> Users { get; }
         string Description { get; }
         bool Opened { get; }
 
@@ -30,7 +30,7 @@ namespace NeoServer.Game.Contracts.Chats
         bool WriteMessage(IPlayer player, string message, out string cancelMessage);
     }
 
-    public interface IChatUser
+    public interface IUserChat
     {
         bool Removed { get; }
         bool IsMuted { get; }
